debundle: remove source-match timing hooks
